Re-Enable Repaint-After-Layout by default.
This CL reenables the new repaint-after-layout feature. This changes the way
that we do repainting in Blink and has the potential to cause under-painting.

If you see underpainting issues, please try with --disable-repaint-after-layout
and see if the issue persists.

Landing on behalf of dsinclair@.

TBR=jchaffraix,dsinclair
BUG=320139

Review URL: https://codereview.chromium.org/263463004

git-svn-id: svn://svn.chromium.org/blink/trunk@172887 bbb929c8-8fbe-4397-9dbb-9b2b20218538