Pass the absolute offset to repaintAfterLayoutIfNeeded
The core issue is that we need to be detect any change in
a renderer's position to trigger a full invalidation. The
old code would detect this case by using the outline box,
which contained the location relative to the repaint
container. The fix is to pass this offset again.

BUG=365289,368024

Review URL: https://codereview.chromium.org/254063002

git-svn-id: svn://svn.chromium.org/blink/trunk@173152 bbb929c8-8fbe-4397-9dbb-9b2b20218538