HTMLAnchorElement.text should not be readonly
HTMLAnchorElement.text should not be readonly as per the HTML specification:
http://www.whatwg.org/specs/web-apps/current-work/multipage/text-level-semantics.html#dom-a-text

On setting HTMLAnchorElement.text, we must act as if the textContent IDL
attribute on the element had been set to the new value. This CL updates Blink
to behave accordingly.

The new behavior is consistent with Firefox 29 and IE11.

R=arv@chromium.org, eseidel@chromium.org
TEST=fast/dom/HTMLAnchorElement/set-text.html
BUG=369950

Review URL: https://codereview.chromium.org/263353004

git-svn-id: svn://svn.chromium.org/blink/trunk@173458 bbb929c8-8fbe-4397-9dbb-9b2b20218538