Fix interaction between --fully-parallel and --run-singly in run-webkit-tests.
By default, with --fully-parallel each test is run in its own shard in
run-webkit-tests, and so a given content_shell may run tests from different
directories in succession; this is normally fine, but when flipping back and
forth between virtual tests and normal tests, you might have to restart
content_shell to change the command line arguments, making things run much
more slowly. So, we keep all of the tests in a virtual suite together
even with --fully-parallel.

*However*, if you're running with --run-singly (or --batch-size=1), like
we do on the valgrind bots, then you are restarting between each test anyway.
Further, we're likely only running a small number of tests at a time,
it's possible that we'd end up with a single shard to run, and not be able
to exploit multiple processes.

This patch changes the logic so that we will also shard every virtual test
if --run-singly is true.

R=ojan@chromium.org
BUG=369325

Review URL: https://codereview.chromium.org/268743007

git-svn-id: svn://svn.chromium.org/blink/trunk@173460 bbb929c8-8fbe-4397-9dbb-9b2b20218538