Make WebWidget::compositeAndReadbackAsync() return void.
This method returns false when isAcceleratedCompositingActive() is
not enabled. In the future this will always be enabled (at least for
callers of the compositeAndReadbackAsync() method), so having to
support a branch and NOTREACHED() at each call site it awkward.

Instead make this method ASSSERT(isAcceleratedCompositingActive()) and
not return a value.

R=abarth
BUG=251960

Review URL: https://codereview.chromium.org/263253002

git-svn-id: svn://svn.chromium.org/blink/trunk@173572 bbb929c8-8fbe-4397-9dbb-9b2b20218538